requirements.txt
streamlit==1.37.1
mysql-connector-python==8.3.0